# docs/swagger.yaml
openapi: 3.0.0
info:
    title: Ajira Apis
    description: Apis for Ajira jobs module
    version: 1.0.0
servers:
    - url: http://localhost:5000/api/v1
      description: Local development server  
paths:
...
    /company:
        post:
            tags:
                - "company"
            summary: Create a company
            description: Restricted to  owners
            operationId: createcompany
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    example: "massive dynamic"
                                description:
                                    type: string
                                    example: "we create everything"
                                logo_url:
                                    type: string
                                    example: "https://storagebucket.gogle.com/massivedy"
                                website_url:
                                    type: string
                                    example: "https://some.massivedynamics.com"
                                email:
                                    type: string
                                    example: "massive.dynamics@info.com"
                            required:
                                - name
                                - description
                                - email
            responses:
                "200":
                    description: Company created successfully, detail of the company are returned
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: integer
                                        example: 1
                                    name:
                                        type: string
                                        example: "massive dynamic"
                                    description:
                                        type: string
                                        example: "we create everything"
                                    logo_url:
                                        type: string
                                        example: "https://storagebucket.gogle.com/massivedy"
                                    owner_id:
                                        type: integer
                                        example: 1
                                    email:
                                        type: string
                                        example: "massive.dynamics@info.com"
                                    active:
                                        type: boolean
                                        example: true
                "500":
                    description: Company name already exists
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: "Name Massive dynamic is already registered"

        get:
            tags:
                - "company"
            parameters:
                - in: query
                  name: queryparams
                  schema:
                      type: object
                      properties:
                          limit:
                              type: integer
                              example: 20
                          nextPage:
                              type: string
                              example: ""
            summary: Get all companies
            description: Restricted to admin accounts
            operationId: getallcompanies
            security:
                - bearerAuth: []
            responses:
                "200":
                    description: An array of objects containing company infos
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        results:
                                            type: array
                                            items:
                                                type: object
                                                properties:
                                                    id:
                                                        type: integer
                                                        example: 1
                                                    name:
                                                        type: string
                                                        example: "massive dynamic"
                                                    description:
                                                        type: string
                                                        example: "we create everything"
                                                    logo_url:
                                                        type: string
                                                        example: "https://storagebucket.gogle.com/massivedy"
                                                    owner_id:
                                                        type: integer
                                                        example: 1
                                                    email:
                                                        type: string
                                                        example: "massive.dynamics@info.com"
                                                    owner:
                                                        type: object
                                                        properties:
                                                            id:
                                                                type: integer
                                                                example: 1
                                                            first_name:
                                                                type: string
                                                                example: "sunday"
                                                            last_name:
                                                                type: string
                                                                example: "oliseh"
                                                            email:
                                                                type: string
                                                                example: "sunday@oliseh.com"
                                                            role:
                                                                type: string
                                                                example: "user"
                                                            phone_number:
                                                                type: string
                                                                example: "0714382366"
                                                            active:
                                                                type: boolean
                                                                example: true
                                                            created_at:
                                                                type: string
                                                                example: "2020-05-05T09:12:57.848Z"
                                                            updated_at:
                                                                type: string
                                                                example: "2020-05-05T09:12:57.848Z"
                                        pageInfo:
                                            type: object
                                            properties:
                                                total:
                                                    type: integer
                                                    example: 30
                                                remaining:
                                                    type: integer
                                                    example: 20
                                                hasMore:
                                                    type: boolean
                                                    example: true
                                                hasPrevious:
                                                    type: boolean
                                                    example: true
                                                hasNext:
                                                    type: boolean
                                                    example: true
                                                next:
                                                    type: string
                                                    example: "KGRhdGUpMjAyMS0wMi0wNFQwMDo1ODoyOS4xODNa"
                                                previous:
                                                    type: string
                                                    example: "KGRhdGUpMjAyMS0wMi0wNFQwMDo1ODoyOS4xODNat"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/UnauthorizedError"
                "404":
                    $ref: "#/components/responses/NotFoundError"
    /company/mine:

        get:
            tags:
                - "company"
            summary: Get myCompany 
            description: Open to all users
            operationId: getmycompany
            security:
                - bearerAuth: []
            responses:
                "200":
                    $ref: "#/components/responses/CompanyResponse"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "404":
                    $ref: "#/components/responses/NotFoundError"                
    /company/{id}:
        parameters:
            - in: path
              name: id
              description: Company ID
              required: true
              example: 1
              schema:
                  type: integer

        get:
            tags:
                - "company"
            summary: Get Company by its ID
            description: Open to all users
            operationId: getcompanybyid
            security:
                - bearerAuth: []
            responses:
                "200":
                    $ref: "#/components/responses/CompanyResponse"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "404":
                    $ref: "#/components/responses/NotFoundError"

        patch:
            tags:
                - "company"
            summary: Update a Company
            description: Restricted to owners and admin
            operationId: updatecompany
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    example: "massive dynamic"
                                description:
                                    type: string
                                    example: "we create everything"
                                logo_url:
                                    type: string
                                    example: "https://storagebucket.gogle.com/massivedy"
                                email:
                                    type: string
                                    example: "massive.dynamics@info.com"
            responses:
                "200":
                    $ref: "#/components/responses/CompanyResponse"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "500":
                    $ref: "#/components/responses/NameExistsError"

        delete:
            tags:
                - "company"
            summary: Delete a Company
            description: Restricted to owners and admin
            operationId: deletecompany
            security:
                - bearerAuth: []
            responses:
                "200":
                    description: Id of the delete company
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: integer
                                        example: 1
                "404":
                    $ref: "#/components/responses/NotFoundError"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
    /company/{id}/company-docs:
        parameters:
            - in: path
              name: id
              description: Company ID
              required: true
              example: 1
              schema:
                  type: integer
                  
        get:
            tags:
                - "company"
            summary: Get Company docs by its ID
            description: Open to all users and admins
            operationId: getcompanydocsbyid
            security:
                - bearerAuth: []
            responses:
                "200":
                    description: Company docs fetched successfully
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        id:
                                            type: integer
                                            example: 1
                                        document_type:
                                            type: string
                                            example: "kra pin"
                                        description:
                                            type: string
                                            example: "kra pin for company x"
                                        document_number:
                                            type: string
                                            example: "A007yrjfbui"
                                        document_url:
                                            type: string
                                            example: "https://some.massivedynamics.com"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "404":
                    $ref: "#/components/responses/NotFoundError"          
        post:
            tags:
                - "company"
            summary: Create a company document
            description: Restricted to  owners
            operationId: createcompanydocument
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                document_type:
                                    type: string
                                    example: "kra pin"
                                description:
                                    type: string
                                    example: "kra pin for company x"
                                document_number:
                                    type: string
                                    example: "A007yrjfbui"
                                document_url:
                                    type: string
                                    example: "https://some.massivedynamics.com"
                                
                            required:
                                - document_type
                                - description
                                - document_url
            responses:
                "200":
                    description: Company created successfully, detail of the company are returned
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: integer
                                        example: 1
                                    document_type:
                                        type: string
                                        example: "kra pin"
                                    description:
                                        type: string
                                        example: "kra pin for company x"
                                    document_number:
                                        type: string
                                        example: "A007yrjfbui"
                                    document_url:
                                        type: string
                                        example: "https://some.massivedynamics.com"
                "500":
                    description: Description is missing
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: "Description is required"                
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    responses:
...
        CompanyResponse:
            description: Company Details
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            id:
                                type: integer
                                example: 1
                            name:
                                type: string
                                example: "massive dynamic"
                            description:
                                type: string
                                example: "we create everything"
                            logo_url:
                                type: string
                                example: "https://storagebucket.gogle.com/massivedy"
                            owner_id:
                                type: integer
                                example: 1
                            email:
                                type: string
                                example: "massive.dynamics@info.com"
                            owner:
                                type: object
                                properties:
                                    id:
                                        type: integer
                                        example: 1
                                    first_name:
                                        type: string
                                        example: "sunday"
                                    last_name:
                                        type: string
                                        example: "oliseh"
                                    email:
                                        type: string
                                        example: "sunday@oliseh.com"
                                    role:
                                        type: string
                                        example: "user"
                                    phone_number:
                                        type: string
                                        example: "0714382366"
                                    active:
                                        type: boolean
                                        example: true
                                    created_at:
                                        type: string
                                        example: "2020-05-05T09:12:57.848Z"
                                    updated_at:
                                        type: string
                                        example: "2020-05-05T09:12:57.848Z"
...
        UnauthorizedError:
            description: Access token is missing or invalid, or the user does not have access to perform the action
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            message:
                                type: string
                                example: "Unauthorized"
        NotFoundError:
            description: Not Found
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            message:
                                type: string
                                example: "Not Found"
...
        NameExistsError:
            description: Name is already in use
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            message:
                                type: string
                                example: "Name is already taken"
...